backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Supabase
supabase==2.5.0

# AI / ML
tensorflow==2.16.1
tf-keras==2.16.0
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80
matplotlib==3.9.0

# RAG / LangChain
langchain==0.2.3
langchain-text-splitters==0.2.0
langchain-google-genai==1.0.6
langchain-community==0.2.3
langchain-chroma==0.1.1
chromadb==0.5.0
google-generativeai==0.7.2

# Embeddings
sentence-transformers==3.0.0
